vector_uint32: push back value

+++ src/vector_uint32.cpp
@@ -12,7 +12,7 @@ void vector_uint32::run()
 {
     for (size_t i = 0; i < m_cycles; ++i)
     {
-        uint16_t value = static_cast<uint16_t>(i);
+        uint32_t value = static_cast<uint32_t>(i);
         m_buffer.push_back(value);
     }
 }
